@@ -22,12 +22,14 @@ namespace QuickApp.ViewModels
             CreateMap<ApplicationUser, UserViewModel>()
                    .ForMember(d => d.Roles, map => map.Ignore());
             CreateMap<UserViewModel, ApplicationUser>()
-                .ForMember(d => d.Roles, map => map.Ignore());
+                .ForMember(d => d.Roles, map => map.Ignore())
+                .ForMember(d => d.Id, map => map.Condition(src => src.Id != null));
 
             CreateMap<ApplicationUser, UserEditViewModel>()
                 .ForMember(d => d.Roles, map => map.Ignore());
             CreateMap<UserEditViewModel, ApplicationUser>()
-                .ForMember(d => d.Roles, map => map.Ignore());
+                .ForMember(d => d.Roles, map => map.Ignore())
+                .ForMember(d => d.Id, map => map.Condition(src => src.Id != null));
 
             CreateMap<ApplicationUser, UserPatchViewModel>()
                 .ReverseMap();
@@ -36,7 +38,8 @@ namespace QuickApp.ViewModels
                 .ForMember(d => d.Permissions, map => map.MapFrom(s => s.Claims))
                 .ForMember(d => d.UsersCount, map => map.MapFrom(s => s.Users != null ? s.Users.Count : 0))
                 .ReverseMap();
-            CreateMap<RoleViewModel, ApplicationRole>();
+            CreateMap<RoleViewModel, ApplicationRole>()
+                .ForMember(d => d.Id, map => map.Condition(src => src.Id != null));
 
             CreateMap<IdentityRoleClaim<string>, ClaimViewModel>()
                 .ForMember(d => d.Type, map => map.MapFrom(s => s.ClaimType))
